+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 0.1.1
+
+Unreleased
+
+- Added no-color option.
+- Fixed to avoid coloring when the output is not terminal.
+
 ## 0.1.0
 
 Released 2021-01-03